backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
yfinance==0.2.40
pandas
numpy
google-genai>=1.0.0
feedparser==6.0.11
python-dotenv==1.0.1
aiohttp==3.9.5
cachetools==5.3.3
python-dateutil==2.9.0
